Renaming API Exception Class
We are renaming the API Exception class to make it easier to work
with in our IDEs. Having multiple files with the same name,
especially files which are related to each other. Such as
src/Exception.php and src/api/Exception.php could be confusing
as most IDEs only who you the filename of the open files unless
you are editing that specific file.