fix: Fixes issue with zip_code vs zipcode
When we switched from using the "Services API" to the new API (api.wolfnet.com) the zip code
argument that the API expects changed from "zipcode" to "zip_code". This change was not
adequately addressed during the migration. This update ensures that zip_code will always be favored,
however if "zipcode" is provided and "zip_code" is no "zipcode" will be translated to "zip_code"
before performing the API request.