feat: Better error message
Adds a better error message for certain errors that may com back from the API. Specifically if a
request comes in that requires the user to be authenticated (double opt-in) we will display a
message specific to that issue.

It is worth noting that there is currently no way for a user to login or authenticate via the
WordPress plugin.

@21773436