feat: Adds a new setting to enable/disable SSL
As a stop gap measure for possible issues that users might have with making SSL requests from their
servers this update adds a new option for enabling and disabling SSL usage by the API Client. The
option will be enabled (SSL on) by default. If any users have issues we can easily disable SSL.
Hopefully no users will have any issues and in the future we will deprecate this feature and
non-SSL support for the API in general.