feat: Updates caching service to remove registry
This update remove the registry concept from the caching service. Instead of adding cache keys to a
registry we instead use a key salt that gets applied to all of the keys that are used to get, set,
and delete values via the Transients API. When we want to invalidate all values in the cache we can
simply change the salt.

Because we are getting rid of the registry we are also updating the way we "reap" (remove expired
values) from the cache. This update reverts us to a method which deletes values directly from the
database. This method only works for the default WP database-based cache. If the site uses a
third-party cache such as memcached or redis the cache system will be responsible for reaping
entries.