refactor: Removes explicit criteria definition
This update removes the explicit definition of criteria which is being sent the API. This means that
any criteria defined in a shortcode will be passed to the API regardless of whether or not they are
used be the API. Not only does this resolve some issues we are experiencing with criteria not being
accounted for, but it also future proofs the criteria.

The only criteria that are hard-coded at this point are done so in a translation capacity. Translating
legacy criteria to newer names or values.

@21827694