fix: Fixes invalid criteria
During refactoring it was missed that the "criteria" array that was being passed to the prepare
method also included a lot of extra data used exclusively for WP widgets. This wouldn't be a
problem except that some of the values were scalar values and not valid for use in the API client.

@21833336