feat: Add multi-market back in
The multi market quick search project is back on the table. This adds back in the ability to
add multiple keys on the Settings page. It also adds new functionality to auto determine the
site to route the quick search to based on which has the most matching listings.

a21881982