feat: Change form POST to Ajax action
Trying to post the quick search and do the processing to determine the correct site to route to
was not working withing the Wordpress plugin architecture. This is due to not being able to
easily use the header function to redirect since headers are already sent by the time the
plugin code runs. So we'll have to use an Ajax request upon form submission instead, and
follow that with a javascript redirect.

a21881982