feat: Add 'View all listings' link
This change requires getting the correct API key associated with an agent in order to make a
request to get their listings. The only way we can do that right now is by adding a market
name to each product key, and then finding the matching key with that market based on
the agent's market. Not a great solution to this problem, but we don't have cross-market
searches yet.

a21901969