fix: Removing code from yesterday and going for a session approach
The approach yesterday of updating the event handlers simply was not working. It was also predicated on the idea that the user
should be able to change the market without clicking the 'Apply' button and then be able to change the primary search type.
That simply would not work because there's no way of knowing if the selected search type even exists in the selected market
until they actually hit 'Apply' on their market selection.

Anyhow, this session-based change is the inherently simpler way to fix the issue at hand, and I probably should have figured
that out sooner...

a21918124