fix: Add instance IDs to javascript so there are no conflicts
Sizing of agent and office boxes had the potential to get out of hand if there were more than one instance of
the agent shortcode at a time (i.e. on the front page with multiple posts, or less likely if there were
two instances of the shortcode in a given post or page, which would be very strange, but hey...).

a21917445